refactor(impl): split up files a bit more
Split up functions into files so that they group better
